renamed function name, moved pytest fixture to conftest.py
